+++ h/app.py
@@ -129,6 +129,7 @@ def includeme(config):
     config.include('h.auth')
     config.include('h.authz')
     config.include('h.db')
+    config.include('h.eventqueue')
     config.include('h.form')
     config.include('h.indexer')
     config.include('h.panels')
